Restructure the lsp module, redirect Sphinx's output to the LSP client
